dataset/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
asttokens==2.4.1
attrs==24.2.0
colorama==0.4.6
colorlog==6.8.2
comm==0.2.2
debugpy==1.8.5
decorator==5.1.1
executing==2.1.0
frozenlist==1.4.1
idna==3.10
ipykernel==6.29.5
ipython==8.27.0
jedi==0.19.1
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
MetaTrader5==5.0.4424
multidict==6.1.0
nest-asyncio==1.6.0
numpy==1.26.4
packaging==24.1
pandas==2.2.3
parso==0.8.4
platformdirs==4.3.6
polars==1.7.1
prompt_toolkit==3.0.47
psutil==6.0.0
psycopg2-binary==2.9.9
pure_eval==0.2.3
pyarrow==17.0.0
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pywin32==306
pyzmq==26.2.0
six==1.16.0
stack-data==0.6.3
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
wcwidth==0.2.13
yarl==1.11.1
